Moving SqlStatement into separate header file
